requirements.txt
# requirements.txt

pandas
langchain
langchain-community
langchain-google-genai
google-generativeai
chromadb
sentence-transformers